#25 refactor: Move database and repository related code into a new project OrderService.Data. All data access is now through the Data project. In the console app using a ServiceCollection to wire up the database context. Passing the ServiceCollection to the repositories when they need a database context.

+++ OrderService.Entities/Order.cs
@@ -6,7 +6,7 @@ int Quantity)
 {
 
   public Guid Id { get; set; }
-  public double Total { get; } = Quantity * Product.Price;
+  public decimal Total { get; } = Quantity * Product.Price;
 
 }
